+++ packages/core/legos/src/sources.ts
@@ -1,6 +1,7 @@
 import * as _1forge from '@chainlink/1forge-adapter'
 import * as accuweather from '@chainlink/accuweather-adapter'
 import * as ada_balance from '@chainlink/ada-balance-adapter'
+import * as alongside from '@chainlink/alongside-adapter'
 import * as alphachain from '@chainlink/alphachain-adapter'
 import * as alphavantage from '@chainlink/alphavantage-adapter'
 import * as alpine from '@chainlink/alpine-adapter'
@@ -19,6 +20,7 @@ import * as blockchain_com from '@chainlink/blockchain.com-adapter'
 import * as blockchair from '@chainlink/blockchair-adapter'
 import * as blockcypher from '@chainlink/blockcypher-adapter'
 import * as blocksize_capital from '@chainlink/blocksize-capital-adapter'
+import * as blocksize_capital_test from '@chainlink/blocksize-capital-test-adapter'
 import * as blockstream from '@chainlink/blockstream-adapter'
 import * as bob from '@chainlink/bob-adapter'
 import * as bravenewcoin from '@chainlink/bravenewcoin-adapter'
@@ -37,6 +39,7 @@ import * as coingecko from '@chainlink/coingecko-adapter'
 import * as coingecko_test from '@chainlink/coingecko-test-adapter'
 import * as coinlore from '@chainlink/coinlore-adapter'
 import * as coinmarketcap from '@chainlink/coinmarketcap-adapter'
+import * as coinmarketcap_test from '@chainlink/coinmarketcap-test-adapter'
 import * as coinmetrics from '@chainlink/coinmetrics-adapter'
 import * as coinmetrics_test from '@chainlink/coinmetrics-test-adapter'
 import * as coinpaprika from '@chainlink/coinpaprika-adapter'
@@ -90,6 +93,7 @@ import * as graphql from '@chainlink/graphql-adapter'
 import * as gsr from '@chainlink/gsr-adapter'
 import * as iex_adapter from '@chainlink/iex-cloud-adapter'
 import * as intrinio from '@chainlink/intrinio-adapter'
+import * as intrinio_test from '@chainlink/intrinio-test-adapter'
 import * as ipfs from '@chainlink/ipfs-adapter'
 import * as jpegd from '@chainlink/jpegd-adapter'
 import * as json_adapter from '@chainlink/json-rpc-adapter'
@@ -105,11 +109,14 @@ import * as marketstack from '@chainlink/marketstack-adapter'
 import * as messari from '@chainlink/messari-adapter'
 import * as metalsapi from '@chainlink/metalsapi-adapter'
 import * as mock_ea from '@chainlink/mock-ea-adapter'
+import * as moonbeam_address_list from '@chainlink/moonbeam-address-list-adapter'
 import * as mycryptoapi from '@chainlink/mycryptoapi-adapter'
 import * as ncfx from '@chainlink/ncfx-adapter'
 import * as nft_blue_chip from '@chainlink/nft-blue-chip-adapter'
 import * as nikkei from '@chainlink/nikkei-adapter'
 import * as nomics from '@chainlink/nomics-adapter'
+import * as nomics_test from '@chainlink/nomics-test-adapter'
+import * as oanda from '@chainlink/oanda-adapter'
 import * as oilpriceapi from '@chainlink/oilpriceapi-adapter'
 import * as onchain_gas from '@chainlink/onchain-gas-adapter'
 import * as openexchangerates from '@chainlink/openexchangerates-adapter'
@@ -117,6 +124,7 @@ import * as orchid_adapter from '@chainlink/orchid-bandwidth-adapter'
 import * as paxos from '@chainlink/paxos-adapter'
 import * as paypal from '@chainlink/paypal-adapter'
 import * as poa from '@chainlink/poa-adapter'
+import * as polkadot_balance from '@chainlink/polkadot-balance-adapter'
 import * as polygon from '@chainlink/polygon-adapter'
 import * as por_address_list from '@chainlink/por-address-list-adapter'
 import * as reduce from '@chainlink/reduce-adapter'
@@ -136,10 +144,13 @@ import * as taapi from '@chainlink/taapi-adapter'
 import * as terra_view_function from '@chainlink/terra-view-function-adapter'
 import * as therundown from '@chainlink/therundown-adapter'
 import * as tiingo from '@chainlink/tiingo-adapter'
+import * as tiingo_test from '@chainlink/tiingo-test-adapter'
 import * as tradermade from '@chainlink/tradermade-adapter'
+import * as tradermade_test from '@chainlink/tradermade-test-adapter'
 import * as tradingeconomics from '@chainlink/tradingeconomics-adapter'
 import * as trueusd from '@chainlink/trueusd-adapter'
 import * as twelvedata from '@chainlink/twelvedata-adapter'
+import * as twosigma from '@chainlink/twosigma-adapter'
 import * as unibit from '@chainlink/unibit-adapter'
 import * as uniswap_v2 from '@chainlink/uniswap-v2-adapter'
 import * as uniswap_v3 from '@chainlink/uniswap-v3-adapter'
@@ -155,6 +166,7 @@ export default {
   _1forge,
   accuweather,
   ada_balance,
+  alongside,
   alphachain,
   alphavantage,
   alpine,
@@ -173,6 +185,7 @@ export default {
   blockchair,
   blockcypher,
   blocksize_capital,
+  blocksize_capital_test,
   blockstream,
   bob,
   bravenewcoin,
@@ -191,6 +204,7 @@ export default {
   coingecko_test,
   coinlore,
   coinmarketcap,
+  coinmarketcap_test,
   coinmetrics,
   coinmetrics_test,
   coinpaprika,
@@ -244,6 +258,7 @@ export default {
   gsr,
   iex_adapter,
   intrinio,
+  intrinio_test,
   ipfs,
   jpegd,
   json_adapter,
@@ -259,11 +274,14 @@ export default {
   messari,
   metalsapi,
   mock_ea,
+  moonbeam_address_list,
   mycryptoapi,
   ncfx,
   nft_blue_chip,
   nikkei,
   nomics,
+  nomics_test,
+  oanda,
   oilpriceapi,
   onchain_gas,
   openexchangerates,
@@ -271,6 +289,7 @@ export default {
   paxos,
   paypal,
   poa,
+  polkadot_balance,
   polygon,
   por_address_list,
   reduce,
@@ -290,10 +309,13 @@ export default {
   terra_view_function,
   therundown,
   tiingo,
+  tiingo_test,
   tradermade,
+  tradermade_test,
   tradingeconomics,
   trueusd,
   twelvedata,
+  twosigma,
   unibit,
   uniswap_v2,
   uniswap_v3,