+++ packages/sources/coingecko-test/src/endpoint/dominance.ts
@@ -1,21 +1,25 @@
 import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
-import { BatchWarmingTransport } from '@chainlink/external-adapter-framework/transports/batch-warming'
 import {
   buildGlobalRequestBody,
   constructEntry,
   GlobalEndpointTypes,
   inputParameters,
 } from '../global-utils'
+import { HttpTransport } from '@chainlink/external-adapter-framework/transports'
+import overrides from '../config/overrides.json'
 
-const batchEndpointTransport = new BatchWarmingTransport<GlobalEndpointTypes>({
-  prepareRequest: (_, config) => buildGlobalRequestBody(config),
+const transport = new HttpTransport<GlobalEndpointTypes>({
+  prepareRequests: (params, config) => buildGlobalRequestBody(params, config),
   parseResponse: (params, res) =>
-    params.map((requestPayload) => constructEntry(res, requestPayload, 'market_cap_percentage')),
+    params.map((requestPayload) =>
+      constructEntry(res.data, requestPayload, 'market_cap_percentage'),
+    ),
 })
 
 export const endpoint = new AdapterEndpoint({
   name: 'dominance',
   aliases: ['market_cap_percentage'],
-  transport: batchEndpointTransport,
+  transport,
   inputParameters,
+  overrides: overrides.coingecko,
 })