+++ packages/sources/coinmetrics-test/src/config/index.ts
@@ -1,4 +1,4 @@
-import { InputParameters } from '@chainlink/external-adapter-framework/validation'
+import { PriceEndpointInputParameters } from '@chainlink/external-adapter-framework/adapter'
 
 // Quote values are used to find a dynamic property in the DP response, in the form of ReferenceRate{quote}
 // Since we use hard property names in response types, we need to make sure that only expected quote values appear in params
@@ -28,21 +28,22 @@ export const customSettings = {
   },
 } as const
 
-export const priceInputParameters: InputParameters = {
-  base: {
-    type: 'string',
-    description: 'The symbol of symbols of the currency to query',
-    required: true,
-    aliases: ['from', 'coin'],
-  },
-  quote: {
-    type: 'string',
-    description: 'The symbol of the currency to convert to',
-    required: true,
-    aliases: ['to', 'market'],
-    options: Object.values(VALID_QUOTES),
-  },
-}
+export const priceInputParameters: PriceEndpointInputParameters & { quote: { options: string[] } } =
+  {
+    base: {
+      type: 'string',
+      description: 'The symbol of symbols of the currency to query',
+      required: true,
+      aliases: ['from', 'coin'],
+    },
+    quote: {
+      type: 'string',
+      description: 'The symbol of the currency to convert to',
+      required: true,
+      aliases: ['to', 'market'],
+      options: Object.values(VALID_QUOTES),
+    },
+  }
 
 export interface ResponseSchema {
   data: {