+++ packages/sources/coinpaprika-test/src/endpoint/crypto.ts
@@ -1,4 +1,4 @@
-import { PriceEndpoint } from '@chainlink/external-adapter-framework/adapter'
+import { CryptoPriceEndpoint } from '@chainlink/external-adapter-framework/adapter'
 import { ProviderResult } from '@chainlink/external-adapter-framework/util'
 import { HttpTransport } from '@chainlink/external-adapter-framework/transports'
 import {
@@ -7,6 +7,7 @@ import {
   EndpointTypes,
   inputParameters,
 } from '../crypto-utils'
+import overrides from '../config/overrides.json'
 
 const httpTransport = new HttpTransport<EndpointTypes>({
   prepareRequests: (params, config) => {
@@ -24,9 +25,10 @@ const httpTransport = new HttpTransport<EndpointTypes>({
   },
 })
 
-export const endpoint = new PriceEndpoint<EndpointTypes>({
+export const endpoint = new CryptoPriceEndpoint<EndpointTypes>({
   name: 'crypto',
   aliases: ['price'],
   transport: httpTransport,
   inputParameters: inputParameters,
+  overrides: overrides.coinpaprika,
 })