+++ packages/sources/cryptocompare-test/src/endpoint/index.ts
@@ -1,1 +1,4 @@
-export { endpoint as cryptoEndpoint } from './crypto'
+export { endpoint as crypto } from './crypto-router'
+export { endpoint as vwap } from './vwap'
+export { endpoint as marketcap } from './marketcap'
+export { endpoint as volume } from './volume'