+++ packages/sources/dxfeed-test/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/dxfeed-test-adapter
 
+## 1.0.2
+
+### Patch Changes
+
+- 65014014d: Bumped v3 framework version. Removed WS_ENABLED env var. Upgraded typescript to version 4.9.5
+
+## 1.0.1
+
+### Patch Changes
+
+- c152ed06e: Bumped v3 framework version
+
 ## 1.0.0
 
 ### Major Changes